requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity==1.14.1
azure-mgmt-compute==30.3.0
azure-mgmt-network==25.1.0
azure-mgmt-resource==23.0.1